Finished splitting out components
