+++ src/components/entity-highlighter/styles.js
@@ -1,0 +1,32 @@
+export const styles = {
+  text: {},
+  highlightText: {
+    color: 'transparent',
+    pointerEvents: 'none',
+    padding: '0',
+    whiteSpace: 'pre-wrap',
+    fontFamily: 'source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace',
+    fontSize: 14,
+  },
+  zeroPos: {
+    textAlign: 'left',
+    position: 'absolute',
+    top: 1,
+    left: 1,
+  },
+  input: {
+    fontFamily: 'source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace',
+    fontSize: 14,
+    background: 'none',
+    border: '1px solid',
+    width: '100%',
+    resize: 'none',
+  },
+  textArea: {
+    position: 'relative', 
+    marginTop: '10px'
+  },
+  entityInput: {
+    marginTop: '20px'
+  }
+};
