+++ src/components/entity-highlighter/utils/hash-string.js
@@ -1,0 +1,10 @@
+export const hashString = (str) => {
+  let hash = 0;
+  if (str.length === 0) return hash;
+  for (let i = 0; i < str.length; i += 1) {
+    const char = str.charCodeAt(i);
+    hash = (hash << 5) - hash + char;
+    hash &= hash; // Convert to 32bit integer
+  }
+  return hash > 0 ? hash : -hash;
+}
